Fixed merge bug, implemented suggestions
