+++ lib/public/components/Detail/index.js
@@ -32,7 +32,7 @@ const format = (postFields, post, selectorName, defaultEmtpy = '') => h(`#${sele
     h('b', `${name}:`),
     (()=> {
         if (format) {
-            return format(post[key]);
+            return format(post[key], post);
         } else if (post[key] !== null) {
             return post[key];
         }
