+++ lib/public/components/RunDetail/index.js
@@ -13,7 +13,8 @@
 
 import format from '../Detail/index.js';
 import { h, iconTrash, iconPlus } from '/js/src/index.js';
-import { formatTimestamp, formatBoolean } from '../../utilities/formatTimestamp.js';
+import { formatTimestamp } from '../../utilities/formatTimestamp.js';
+import { formatBoolean } from '../../utilities/formatBoolean.js';
 import { formatRunDuration } from '../../views/Runs/formatRunDuration.js';
 
 /**
@@ -103,11 +104,11 @@ const activeFields = (model) => ({
     trgGlobalRunEnabled: {
         name: 'Trigger Global Run Enabled',
         visible: true,
-        format: (boolean) => formatBoolean(boolean),
+        format: formatBoolean,
     },
     trgEnabled: {
         name: 'Trigger Enabled',
-        format: (boolean) => formatBoolean(boolean),
+        format: formatBoolean,
     },
     pdpConfigOption: {
         name: 'PDP Configuration Option',
@@ -142,19 +143,19 @@ const activeFields = (model) => ({
         name: 'Data Distribution (FLP)',
         visible: true,
         size: 'cell-s',
-        format: (boolean) => formatBoolean(boolean),
+        format: formatBoolean,
     },
     dcs: {
         name: 'DCS',
         visible: true,
         size: 'cell-s',
-        format: (boolean) => formatBoolean(boolean),
+        format: formatBoolean,
     },
     epn: {
         name: 'EPN',
         visible: true,
         size: 'cell-s',
-        format: (boolean) => formatBoolean(boolean),
+        format: formatBoolean,
     },
     epnTopology: {
         name: 'EPN Topology',
