+++ lib/public/utilities/formatTimestamp.js
@@ -43,21 +43,3 @@ export const formatTimestamp = (timestamp, inline = true) => {
         date.toLocaleTimeString(LOCALE_FORMAT, { timeStyle: TIME_STYLE }),
     ]);
 };
-
-/**
- * Returns a textual value by the use of a boolean or a null value.
- *
- * @param {boolean|null} boolean The given boolean
- *
- * @return {string} If present it returs on or off, if not it returns a dash.
- *
- */
-export const formatBoolean = (boolean) => {
-    if (boolean != null) {
-        if (boolean) {
-            return 'On';
-        }
-        return 'Off';
-    }
-    return '-';
-};
